made authorised requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,15 +1,25 @@
 import {useWorkoutsContext} from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const {dispatch}= useWorkoutsContext()
+  const {user} = useAuthContext()
 
   const handleClick = async () => {
+
+    if (!user) {
+      return
+    }
+
     const response = await fetch(
       `${process.env.REACT_APP_API_URL}/api/workouts/${workout._id}`, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
@@ -19,6 +29,9 @@ const WorkoutDetails = ({ workout }) => {
   }
 
   const handleChange = async () => {
+    // if (!user) {
+    //   return
+    // }
     // e.preventDefault()
 
     // const workout = {title, load, reps}
@@ -26,6 +39,9 @@ const WorkoutDetails = ({ workout }) => {
     const response = await fetch(
       `${process.env.REACT_APP_API_URL}/api/workouts/${workout._id}`, {
       method: 'PATCH',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
       // body: JSON.stringify(workout),
       // headers: {
       //   'Content-Type' : 'application/json',
